+++ crates/omnix-ci/crate.nix
@@ -10,9 +10,7 @@ let
   inherit (inputs) self;
 in
 {
-  autoWire = lib.optionals
-    (lib.elem pkgs.system [ "x86_64-linux" "aarch64-darwin" ])
-    [ "doc" "clippy" ];
+  autoWire = [ ];
   crane = {
     args = {
       nativeBuildInputs = with pkgs; with pkgs.apple_sdk_frameworks; lib.optionals stdenv.isDarwin [
@@ -38,6 +36,9 @@ in
         DEFAULT_FLAKE_SCHEMAS
         INSPECT_FLAKE
         NIX_SYSTEMS
+        ;
+      inherit (rust-project.crates."omnix-health".crane.args)
+        CACHIX_BIN
         ;
       DEVOUR_FLAKE = inputs.devour-flake;