+++ crates/omnix-develop/crate.nix
@@ -1,14 +1,11 @@
-{ flake
-, pkgs
+{ pkgs
 , lib
 , rust-project
 , ...
 }:
 
 {
-  autoWire = lib.optionals
-    (lib.elem pkgs.system [ "x86_64-linux" "aarch64-darwin" ])
-    [ "doc" "clippy" ];
+  autoWire = [ ];
   crane.args = {
     buildInputs = lib.optionals pkgs.stdenv.isDarwin (
       with pkgs.apple_sdk_frameworks; [
@@ -19,6 +16,9 @@
       DEFAULT_FLAKE_SCHEMAS
       INSPECT_FLAKE
       NIX_SYSTEMS
+      ;
+    inherit (rust-project.crates."omnix-health".crane.args)
+      CACHIX_BIN
       ;
   };
 }